Updating support library version

+++ app/src/main/java/acr/browser/lightning/constant/Proxy.kt
@@ -8,6 +8,6 @@ import android.support.annotation.IntDef
  *
  * These should match the order of @array/proxy_choices_array
  */
-@IntDef(NO_PROXY.toLong(), PROXY_ORBOT.toLong(), PROXY_I2P.toLong(), PROXY_MANUAL.toLong())
+@IntDef(NO_PROXY, PROXY_ORBOT, PROXY_I2P, PROXY_MANUAL)
 @Retention(AnnotationRetention.SOURCE)
 annotation class Proxy
